# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/genetic-data": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List Genetic Data records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeneticData"
                ],
                "summary": "List Genetic Data records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "doctor_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Analysis Date",
                        "name": "analysis_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.GetAllGeneticDateRes"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new Genetic Data record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeneticData"
                ],
                "summary": "Create a new Genetic Data record",
                "parameters": [
                    {
                        "description": "body for GeneticDataCreate",
                        "name": "geneticdata",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.GeneticDataCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/genetic-data/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Genetic Data by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeneticData"
                ],
                "summary": "Get Genetic Data by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Genetic Data ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.GeneticDataRes"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Genetic Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeneticData"
                ],
                "summary": "Update Genetic Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Genetic Data ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "body for GeneticDataUpdate",
                        "name": "geneticdata",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.GeneticDataUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Genetic Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeneticData"
                ],
                "summary": "Delete Genetic Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Genetic Data ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "health_sync.GeneticDataCreate": {
            "type": "object",
            "properties": {
                "analysis_date": {
                    "type": "string"
                },
                "data_type": {
                    "type": "string"
                },
                "data_value": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "health_sync.GeneticDataRes": {
            "type": "object",
            "properties": {
                "analysis_date": {
                    "type": "string"
                },
                "data_type": {
                    "type": "string"
                },
                "data_value": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "health_sync.GeneticDataUpdate": {
            "type": "object",
            "properties": {
                "analysis_date": {
                    "type": "string"
                },
                "data_type": {
                    "type": "string"
                },
                "data_value": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "health_sync.GetAllGeneticDateRes": {
            "type": "object",
            "properties": {
                "GeneticDatas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/health_sync.GeneticDataRes"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}